refactor: consistant test folder name
