+++ test/fixtures/removing-attrs-to-be-removed.json
