+++ test/fixtures/removing-tags-to-be-removed.json
